Added cupertino styled alert

+++ lib/screens/manage_products_screen.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
@@ -10,14 +11,49 @@ import '../widgets/drawer.dart';
 
 class ManageProductsScreen extends StatelessWidget {
   static const routeName = '/manage_products';
+
+  Future<bool> _showAlert(BuildContext ctx) {
+    return showCupertinoDialog(
+      context: ctx,
+      builder: (_) {
+        return CupertinoAlertDialog(
+          title: Text('Do you really want to delete this?'),
+          actions: <Widget>[
+            CupertinoDialogAction(
+              child: FlatButton(
+                onPressed: () {
+                  Navigator.of(ctx).pop(true);
+                },
+                child: Text('YES'),
+              ),
+            ),
+            CupertinoDialogAction(
+              child: FlatButton(
+                onPressed: () {
+                  Navigator.of(ctx).pop(false);
+                },
+                child: Text('NO'),
+              ),
+            ),
+          ],
+        );
+      },
+    );
+  }
+
+  _deleteProduct(BuildContext ctx, ProductProvider productProvider, int index) {
+    _showAlert(ctx).then(
+      (isOkayToDelete) {
+        if (isOkayToDelete) {
+          return productProvider.deleteProduct(productProvider.items[index].id);
+        }
+      },
+    );
+  }
+
   @override
   Widget build(BuildContext context) {
     final productProvider = Provider.of<ProductProvider>(context);
-
-    Function _deleteProduct(int index) {
-      print('index: $index');
-      return productProvider.deleteProduct(productProvider.items[index].id);
-    }
 
     return Scaffold(
       appBar: AppBar(
@@ -42,7 +78,7 @@ class ManageProductsScreen extends StatelessWidget {
                   productProvider.items[index].id,
                   productProvider.items[index].title,
                   productProvider.items[index].imageUrl,
-                  () => _deleteProduct(index),
+                  () => _deleteProduct(context, productProvider, index),
                 ),
                 Divider(color: Theme.of(context).textTheme.body1.color),
               ],
